AI/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
facenet-pytorch==2.5.3
fastapi==0.118.0
filelock==3.19.1
filetype==1.2.0
fsspec==2025.9.0
google-ai-generativelanguage==0.8.0
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.0
langchain-core==1.0.0
langchain-google-genai==3.0.0
langchain-openai==1.0.1
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
openai==2.6.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tools==1.0.6
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
xxhash==3.6.0
zstandard==0.25.0
